move operation rule code to class
