+++ src/peersafe/app/tx/TableListSet.h
@@ -32,12 +32,6 @@ namespace ripple {
 class TableListSet
     : public ChainSqlTx
 {
-    /* The largest number of paths we allow */
-    static std::size_t const MaxPathSize = 6;
-
-    /* The longest path we allow */
-    static std::size_t const MaxPathLength = 8;
-
 public:
     TableListSet(ApplyContext& ctx)
         : ChainSqlTx(ctx)
@@ -64,8 +58,6 @@ public:
     TER doApply () override;
 
 	TER dealWithOperationRule(ApplyView& view, const STTx & tx, Application& app);
-
-	TER preApply();
 
     static
     STObject generateTableEntry(const STTx &tx, ApplyView& view);
