+++ src/peersafe/rpc/TableUtils.h
@@ -27,7 +27,9 @@
 namespace ripple {
 
 	Json::Value generateError(const std::string& errMsg, bool ws = false);
-	STEntry * getTableEntry(const STArray & aTables, std::string sCheckName);
+	STEntry * getTableEntry(const STArray & aTables, std::string sCheckName); 
+	STEntry *getTableEntry(ApplyView& view, const STTx& tx);
+	STEntry *getTableEntry(const STArray & aTables, Blob& vCheckName);
 	bool isChainSqlBaseType(const std::string& transactionType);
 }
