+++ src/ripple/rpc/impl/TransactionSign.cpp
@@ -40,7 +40,7 @@
 #include <ripple/rpc/impl/LegacyPathFind.h>
 #include <ripple/rpc/impl/RPCHelpers.h>
 #include <ripple/rpc/impl/Tuning.h>
-#include <peersafe/rpc/impl/TableUtils.h>
+#include <peersafe/rpc/TableUtils.h>
 #include <algorithm>
 #include <iterator>
