+++ src/ripple/unity/app_tx.cpp
@@ -44,3 +44,4 @@
 #include <peersafe/app/tx/impl/TableListSet.cpp>
 #include <peersafe/app/tx/impl/SqlStatement.cpp>
 #include <peersafe/app/tx/impl/SqlTransaction.cpp>
+#include <peersafe/app/tx/impl/OperationRule.cpp>
